Pulse v1 Half done
Factories can now produce goods based on the amount of resources given
to them. Sectors distribute and recover the goods. It hasn't been
streamlined or tested so far; the streamlining will wait for a complete
version, since there's no point in optimizing an algorithm that I'm not
going to use. The testing will occur when the system is complete, since
it shouldn't take too long to implement the next part of Pulse v1.