GUI outputting useful information
Clicking on a sector now gives a breakdown of all resources produced and
consumed by that sector. Running into a bug where the sector buttons
fail to appear sometimes, but resizing or restarting the program seems
to fix it. Some more work needs to be done for the formatting of the
information produced; I've hard-coded the distances and left empty
spaces where a product is consumed but not produced or vice versa. Also,
the text area looks ugly, but I don't know if that's worth fixing at the
moment. Still no graphics for sector connections, and economic
recommendations will have to wait for V2 of the product. Next milestones
to tackle are actual economic simulation, which requires the data entry
of all the resource, factory, and sector data, or updating the graphics
to be less of an eyesore.