Factories added to database
Had to bodge things a little to smooth the transition (the ending to the
resource file is like that to avoid unnecessary if checks), but all of
the factories are in and they should be linked to wares correctly. I
recorded the race-specific wares and ware amounts with -1; that needs to
be checked in factory instantiation.